+++ testutil-server/src/test/java/io/spine/server/procman/ProcessManagerCommandTestShould.java
@@ -40,7 +40,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
  */
 @SuppressWarnings("DuplicateStringLiteralInspection")
 @DisplayName("ProcessManagerCommandTest should")
-class ProcessManagerCommandTestTest {
+class ProcessManagerCommandTestShould {
 
     private TimestampProcessManagerTest pmCommandTest;
